refactor(kafka): keep kafka.go layout and isolate deferred-ack types
Move write/read/config back into kafka.go so the PR delta against main stays small, and leave watermark tracking plus mutex-wrapped consumer calls in offsets.go and reader.go.
